requirements.txt
torch
torchvision
torchaudio
transformers
scikit-learn
numpy
pillow
tqdm
joblib
opencv-python
flask
requests
python-dotenv
bcrypt
PyJWT
google-auth
cloudinary
fastapi
uvicorn
python-multipart
faiss-cpu
av
sqlalchemy
apscheduler
psycopg2-binary
pytest
